requirements.txt
requests
yara-python